+++ R/cdm.R
@@ -1,7 +1,7 @@
 
 #' Create a CDM reference object from a database connection
 #'
-#' @param con A DBI database connection to a database where an OMOP CDM v5.4 instance is located.
+#' @param con A DBI database connection to a database where an OMOP CDM v5.4 or v5.3 instance is located.
 #' @param cdm_schema The schema where the OMOP CDM tables are located. Defaults to NULL.
 #' @param write_schema An optional schema in the CDM database that the user has write access to.
 #' @param cdm_tables Which tables should be included? Supports a character vector, tidyselect selection helpers, or table groups.
@@ -11,19 +11,34 @@
 #'   \item{tbl_group("clinical")}{the clinical CDM tables}
 #' }
 #' @param cohort_tables A character vector listing the cohort table names to be included in the CDM object.
+#' @param cdm_version The version of the OMOP CDM: "5.3" (default), "5.4", "auto".
+#' "auto" attempts to automatically determine the cdm version using heuristics.
 #' Cohort tables must be in the write_schema.
 #' @return A list of dplyr database table references pointing to CDM tables
 #' @importFrom dplyr all_of matches starts_with ends_with contains
 #' @export
-cdm_from_con <- function(con, cdm_schema = NULL, cdm_tables = tbl_group("default"), write_schema = NULL, cohort_tables = NULL) {
+cdm_from_con <- function(con, cdm_schema = NULL, cdm_tables = tbl_group("default"), write_schema = NULL, cohort_tables = NULL, cdm_version = "5.3") {
   checkmate::assert_class(con, "DBIConnection")
   checkmate::assert_true(DBI::dbIsValid(con))
   checkmate::assert_character(cdm_schema, null.ok = TRUE, min.len = 1, max.len = 2)
   checkmate::assert_character(write_schema, null.ok = TRUE, min.len = 1, max.len = 2)
   checkmate::assert_character(cohort_tables, null.ok = TRUE, min.len = 1)
+  checkmate::assert_choice(cdm_version, choices = c("5.3", "5.4", "auto"))
+
+  # handle schema names like 'schema.dbo'
+  if (!is.null(cdm_schema) && length(cdm_schema) == 1) {
+    cdm_schema <- strsplit(cdm_schema, "\\.")[[1]]
+    checkmate::assert_character(cdm_schema, null.ok = TRUE, min.len = 1, max.len = 2)
+  }
+  if (!is.null(write_schema) && length(write_schema) == 1) {
+    write_schema <- strsplit(write_schema, "\\.")[[1]]
+    checkmate::assert_character(write_schema, null.ok = TRUE, min.len = 1, max.len = 2)
+  }
+
+  if (cdm_version == "auto") cdm_version <- detect_cdm_version(con, cdm_schema = cdm_schema)
 
   # tidyselect: https://tidyselect.r-lib.org/articles/tidyselect.html
-  all_cdm_tables <- rlang::set_names(spec_cdm_table$cdmTableName, spec_cdm_table$cdmTableName)
+  all_cdm_tables <- rlang::set_names(spec_cdm_table[[cdm_version]]$cdmTableName, spec_cdm_table[[cdm_version]]$cdmTableName)
   cdm_tables <- names(tidyselect::eval_select(rlang::enquo(cdm_tables), data = all_cdm_tables))
 
   if (dbms(con) == "duckdb") {
@@ -57,7 +72,85 @@ cdm_from_con <- function(con, cdm_schema = NULL, cdm_tables = tbl_group("default
   attr(cdm, "cdm_schema") <- cdm_schema
   attr(cdm, "write_schema") <- write_schema
   attr(cdm, "dbcon") <- con
+  attr(cdm, "cdm_version") <- cdm_version
   cdm
+}
+
+detect_cdm_version <- function(con, cdm_schema = NULL) {
+
+  cdm_tables <- c("visit_occurrence", "cdm_source", "procedure_occurrence")
+  if (!all(cdm_tables %in% listTables(con, schema = cdm_schema))) {
+    rlang::abort(paste0("The ", paste(cdm_tables, collapse = ", "), " tables are require for auto-detection of cdm version."))
+  }
+  if (dbms(con) == "duckdb") {
+    cdm <- purrr::map(cdm_tables, ~dplyr::tbl(con, paste(c(cdm_schema, .), collapse = ".")))
+  } else if (is.null(cdm_schema)) {
+    cdm <- purrr::map(cdm_tables, ~dplyr::tbl(con, .))
+  } else if (length(cdm_schema) == 1) {
+    cdm <- purrr::map(cdm_tables, ~dplyr::tbl(con, dbplyr::in_schema(cdm_schema, .)))
+  } else if (length(cdm_schema) == 2) {
+    cdm <- purrr::map(cdm_tables, ~dplyr::tbl(con, dbplyr::in_catalog(cdm_schema[1], cdm_schema[2], .)))
+  }
+  names(cdm) <- cdm_tables
+
+  # Try a few different things to figure out what the cdm version is
+  visit_occurrence_names <- cdm$visit_occurrence %>% head() %>% collect() %>% names() %>% tolower()
+  if ("admitting_source_concept_id" %in% visit_occurrence_names) return("5.3")
+  if ("admitted_from_concept_id" %in% visit_occurrence_names) return("5.4")
+
+  procedure_occurrence_names <- cdm$procedure_occurrence %>% head() %>% collect() %>% names() %>% tolower()
+  if ("procedure_end_date" %in% procedure_occurrence_names) return("5.4")
+
+  cdm_version <- cdm$cdm_source %>% dplyr::pull(.data$cdm_version)
+  if (isTRUE(grepl("5\\.4", cdm_version))) return("5.4")
+  if (isTRUE(grepl("5\\.3", cdm_version))) return("5.3")
+
+  if ("episode" %in% listTables(con, schema = cdm_schema)) return("5.4") else return("5.3")
+
+  # rlang::abort("cdm version could not be automatically detected.")
+}
+
+#' Get the CDM version
+#'
+#' @param cdm A cdm object
+#'
+#' @return "5.3" or "5.4"
+#' @export
+#'
+#' @examples
+#' \dontrun{
+#' library(CDMConnector)
+#' con <- DBI::dbConnect(duckdb::duckdb(), eunomia_dir())
+#' cdm <- cdm_from_con(con, cdm_tables = c(tbl_group("default"), "cdm_source"))
+#' version(cdm)
+#'
+#' DBI::dbDisconnect(con, shutdown = TRUE)
+#' }
+version <- function(cdm) {
+  checkmate::assert_class(cdm, "cdm_reference")
+  v <- attr(cdm, "cdm_version")
+  if (!(v %in% c("5.3", "5.4"))) rlang::abort("cdm object version attribute is not 5.3 or 5.4. Contact the maintainer.")
+  v
+}
+
+#' Create a CDM reference object from a database connection
+#'
+#' @param con A DBI database connection to a database where an OMOP CDM v5.4 instance is located.
+#' @param cdmSchema The schema where the OMOP CDM tables are located. Defaults to NULL.
+#' @param writeSchema An optional schema in the CDM database that the user has write access to.
+#' @param cdmTables Which tables should be included? Supports a character vector, tidyselect selection helpers, or table groups.
+#' \itemize{
+#'   \item{tbl_group("all")}{all CDM tables}
+#'   \item{tbl_group("vocab")}{the CDM vocabulary tables}
+#'   \item{tbl_group("clinical")}{the clinical CDM tables}
+#' }
+#' @param cohortTables A character vector listing the cohort table names to be included in the CDM object.
+#' Cohort tables must be in the write_schema.
+#' @return A list of dplyr database table references pointing to CDM tables
+#' @importFrom dplyr all_of matches starts_with ends_with contains
+#' @export
+cdmFromCon <- function(con, cdmSchema = NULL, cdmTables = tbl_group("default"), writeSchema = NULL, cohortTables = NULL) {
+  cdm_from_con(con = con, cdm_schema = cdmSchema, cdm_tables = cdmTables, write_schema = writeSchema, cohort_tables = cohortTables)
 }
 
 #' Print a CDM reference object
@@ -76,15 +169,40 @@ print.cdm_reference <- function(x, ...) {
   invisible(x)
 }
 
-verify_write_access <- function(con, write_schema) {
+# con = database connection
+# write_schema = schema with write access
+# add = checkmate collection
+verify_write_access <- function(con, write_schema, add = NULL) {
+  checkmate::assert_character(write_schema, min.len = 1, max.len = 2, min.chars = 1)
+  checkmate::assert_class(add, "AssertCollection", null.ok = TRUE)
+  checkmate::assert_true(DBI::dbIsValid(con))
+
   write_schema <- paste(write_schema, collapse = ".")
   tablename <- paste(c(sample(letters, 12, replace = TRUE), "_test_table"), collapse = "")
   tablename <- paste(write_schema, tablename, sep = ".")
-  # spec_cdm_table is a global internal package dataframe
-  DBI::dbWriteTable(con, DBI::SQL(tablename), spec_cdm_table[1:4,])
-  to_compare <- DBI::dbReadTable(con, DBI::SQL(tablename))
+
+  df1 <- data.frame(chr_col = "a", numeric_col = 1, int_col = 1L)
+  # df1 <- data.frame(chr_col = "a", dbl_col = 1.0)
+  DBI::dbWriteTable(con, DBI::SQL(tablename), df1)
+  withr::with_options(list(databaseConnectorIntegerAsNumeric = FALSE), {
+    df2 <- DBI::dbReadTable(con, DBI::SQL(tablename))
+  })
   DBI::dbRemoveTable(con, DBI::SQL(tablename))
-  if(!dplyr::all_equal(spec_cdm_table[1:4,], to_compare)) rlang::abort(paste("Write access to schema", write_schema, "could not be verified."))
+
+  # if(is(con, "DatabaseConnectorConnection") && isTRUE(dplyr::all_equal(df1[,-2], df2[,-2]))) {
+  #   rlang::inform("Write access verified. Note TRUE/FALSE values and upper case table names are not supported by `DatabaseConnector`.",
+  #                 .frequency = "once", .frequency_id = "DatabaseConnectorBooleans")
+  #   if (is.null(add)) rlang::abort(msg) else add$push(msg)
+  #
+  # } else if(!isTRUE(dplyr::all_equal(df1, df2))) {
+  #   msg <- paste("Write access to schema", write_schema, "could not be verified.")
+  #   if (is.null(add)) rlang::abort(msg) else add$push(msg)
+  # }
+
+  if(!isTRUE(dplyr::all_equal(df1, df2))) {
+    msg <- paste("Write access to schema", write_schema, "could not be verified.")
+    if (is.null(add)) rlang::abort(msg) else add$push(msg)
+  }
   invisible(NULL)
 }
 
@@ -101,7 +219,7 @@ verify_write_access <- function(con, write_schema) {
 #' device_exposure, measurement, observation, death, note, note_nlp, specimen,
 #' fact_relationship, location, care_site, provider, payer_plan_period,
 #' cost, drug_era, dose_era, condition_era, concept, vocabulary, concept_relationship,
-#' concept_ancestor, drug_strength
+#' concept_ancestor, concept_synonym, drug_strength
 #'
 #' @param group A character vector of CDM table groups: "vocab", "clinical", "all", "default", "derived".
 #'
@@ -121,7 +239,8 @@ verify_write_access <- function(con, write_schema) {
 tbl_group <- function(group) {
   # groups are defined in the internal package dataframe called spec_cdm_table created by a script in the extras folder
   checkmate::assert_subset(group, c("vocab", "clinical", "all", "default", "derived"))
-  purrr::map(group, ~spec_cdm_table[spec_cdm_table[[paste0("group_", .)]], ]$cdmTableName) %>% unlist() %>% unique()
+  spec <- spec_cdm_table[["5.3"]] # use v5.3 here. Currently the set of table groups between 5.3 and 5.4 are the same.
+  purrr::map(group, ~spec[spec[[paste0("group_", .)]], ]$cdmTableName) %>% unlist() %>% unique()
 }
 
 #' Create a new Eunomia CDM
@@ -133,7 +252,7 @@ tbl_group <- function(group) {
 #'
 #' @return The full path to the new Eunomia CDM that can be passed to `dbConnect()`
 #' @export
-#' @importFrom utils untar
+#' @importFrom utils untar download.file menu
 #' @examples
 #' \dontrun{
 #' library(DBI)
@@ -143,15 +262,51 @@ tbl_group <- function(group) {
 #' dbDisconnect(con)
 #' }
 eunomia_dir <- function(exdir = NULL) {
+  rlang::check_installed("duckdb", version = "0.6.0", reason = "duckdb version 0.6 is required to use eunomia_dir()")
+  if (!eunomia_is_available() && interactive()) {
+    rc <- menu(c("Yes", "No"),
+              title = "Eunomia has not been downloaded yet.\nWould you like to download the Eunomia example CDM dataset?")
+  } else {
+    rc <- 1 # Automatically try downloading Eunomia if R is running non-interactively.
+  }
+
+  if (rc != 1) rlang::abort("Eunomia dataset is not available")
+
   if (is.null(exdir)) exdir <- file.path(tempdir(TRUE), paste(sample(letters, 8, replace = TRUE), collapse = ""))
-  file <- xzfile(system.file("duckdb", "cdm.duckdb.tar.xz", package = "CDMConnector"), open = "rb")
+  file <- xzfile(eunomia_cache(), open = "rb")
   untar(file, exdir = exdir)
   close(file)
   path <- file.path(exdir, "cdm.duckdb")
   if(!file.exists(path)) rlang::abort("Error creating Eunomia CDM")
-  path
+  return(path)
 }
 
+# has the eunomia dataset been cached?
+eunomia_is_available <- function() {
+  path <- Sys.getenv("EUNOMIA_CACHE_DIR", unset = "~/eunomia_cache")
+  path <- path.expand(path)
+  file.exists(file.path(path, "cdm.duckdb.tar.xz"))
+}
+
+# return the location to the cached eunomia dataset
+eunomia_cache <- function() {
+  path <- Sys.getenv("EUNOMIA_CACHE_DIR", unset = "~/eunomia_cache")
+  path <- path.expand(path)
+  if (!eunomia_is_available()) {
+    if (!dir.exists(path)) dir.create(path)
+
+    download.file("https://github.com/OdyOSG/EunomiaData/raw/main/cdm.duckdb.tar.xz",
+                  file.path(path, "cdm.duckdb.tar.xz"),
+                  method = "auto")
+  }
+  if (!eunomia_is_available()) rlang::abort("Error downloading Eunomia dataset.")
+
+  return(file.path(path, "cdm.duckdb.tar.xz"))
+}
+
+#' @export
+#' @describeIn eunomia_dir camelCase alias
+eunomiaDir <- eunomia_dir
 
 #' Get the database management system (dbms) from a cdm_reference or DBI connection
 #'
@@ -226,11 +381,20 @@ stow <- function(cdm, path, format = "parquet") {
 #' @param format What is the file format to be read in? Must be "auto" (default), "parquet", "csv", "feather".
 #' @param as_data_frame TRUE (default) will read files into R as dataframes. FALSE will read files into R as Arrow Datasets.
 #' @return A list of dplyr database table references pointing to CDM tables
+#' @importFrom lifecycle deprecated
 #' @export
-cdm_from_files <- function(path, cdm_tables = tbl_group("default"), format = "auto", as_data_frame = TRUE) {
+cdm_from_files <- function(path, cdm_tables = deprecated(), format = "auto", as_data_frame = TRUE) {
   checkmate::assert_choice(format, c("auto", "parquet", "csv", "feather"))
+  checkmate::assert_logical(as_data_frame, len = 1, null.ok = FALSE)
   checkmate::assert_true(file.exists(path))
 
+  if (lifecycle::is_present(cdm_tables)) {
+    lifecycle::deprecate_warn(
+      when = "0.2.0",
+      what = "cdm_from_files(cdm_tables)",
+      details = "Ability to select a subset of cdm tables when reading a cdm from files has been deprecated. All cdm table files are read into R."
+    )
+  }
   path <- path.expand(path)
 
   files <- list.files(path, full.names = TRUE)
@@ -241,26 +405,38 @@ cdm_from_files <- function(path, cdm_tables = tbl_group("default"), format = "au
     checkmate::assert_choice(format, c("parquet", "csv", "feather"))
   }
 
-  # tidyselect: https://tidyselect.r-lib.org/articles/tidyselect.html
-  all_cdm_tables <- rlang::set_names(spec_cdm_table$cdmTableName, spec_cdm_table$cdmTableName)
-  cdm_tables <- names(tidyselect::eval_select(rlang::enquo(cdm_tables), data = all_cdm_tables))
-
+  cdm_tables <- tools::file_path_sans_ext(basename(list.files(path)))
   cdm_table_files <- file.path(path, paste0(cdm_tables, ".", format))
   purrr::walk(cdm_table_files, function(.) checkmate::assert_file_exists(., "r"))
 
   cdm <- switch (format,
-    parquet = purrr::map(cdm_table_files, function(.) arrow::read_parquet(., as_data_frame = as_data_frame)),
-    csv = purrr::map(cdm_table_files, function(.) arrow::read_csv_arrow(., as_data_frame = as_data_frame)),
-    feather = purrr::map(cdm_table_files, function(.) arrow::read_feather(., as_data_frame = as_data_frame))
+                 parquet = purrr::map(cdm_table_files, function(.) arrow::read_parquet(., as_data_frame = as_data_frame)),
+                 csv = purrr::map(cdm_table_files, function(.) arrow::read_csv_arrow(., as_data_frame = as_data_frame)),
+                 feather = purrr::map(cdm_table_files, function(.) arrow::read_feather(., as_data_frame = as_data_frame))
   ) %>%
-  magrittr::set_names(cdm_tables) %>%
-  magrittr::set_class("cdm_reference")
+    magrittr::set_names(cdm_tables) %>%
+    magrittr::set_class("cdm_reference")
 
   attr(cdm, "cdm_schema") <- NULL
   attr(cdm, "write_schema") <- NULL
   attr(cdm, "dbcon") <- NULL
+  attr(cdm, "cdm_version") <- NULL
   cdm
 }
+
+#' Create a CDM reference from a folder containing parquet, csv, or feather files
+#'
+#' @param path A folder where an OMOP CDM v5.4 instance is located.
+#' @param cdmTables Which tables should be included? Supports tidyselect and custom selection groups.
+#' @param format What is the file format to be read in? Must be "auto" (default), "parquet", "csv", "feather".
+#' @param as_data_frame TRUE (default) will read files into R as dataframes. FALSE will read files into R as Arrow Datasets.
+#' @return A list of dplyr database table references pointing to CDM tables
+#' @export
+cdmFromFiles <- function(path, cdmTables = tbl_group("default"), format = "auto", as_data_frame = TRUE) {
+  cdm_from_files(path = path, cdm_tables = cdmTables, format = format, as_data_frame = as_data_frame)
+}
+
+# collect <- function(x, ...) {UseMethod("collect")}
 
 #' Bring a remote CDM reference into R
 #'
@@ -287,3 +463,57 @@ collect.cdm_reference <- function(x, ...) {
   }
   x
 }
+
+##' @importFrom dplyr collect
+##' @export
+NULL
+
+#' Extract CDM metadata
+#'
+#' Extract the name, version, and selected record counts from a cdm.
+#'
+#' @param cdm A cdm object
+#'
+#' @return A list of attributes about the cdm including selected fields from the
+#' cdm_source table and record counts from the person and observation_period tables
+#' @export
+#'
+#' @examples
+#' \dontrun{
+#' library(CDMConnector)
+#' con <- DBI::dbConnect(duckdb::duckdb(), eunomia_dir())
+#' cdm <- cdm_from_con(con, cdm_tables = c(tbl_group("default"), "cdm_source"))
+#' snapshot(cdm)
+#'
+#' DBI::dbDisconnect(con, shutdown = TRUE)
+#' }
+snapshot <- function(cdm) {
+
+  assert_tables(cdm, tables = c("cdm_source", "person", "observation_period", "vocabulary"))
+
+  person_cnt <- dplyr::tally(cdm$person, name = "n") %>% dplyr::pull(.data$n)
+  observation_period_cnt <- dplyr::tally(cdm$observation_period, name = "n") %>% dplyr::pull(.data$n)
+  vocab_version <- cdm$vocabulary %>% dplyr::filter(.data$vocabulary_id == "None") %>% dplyr::pull(.data$vocabulary_version)
+  cdm_source_name <- cdm$cdm_source %>% dplyr::pull(.data$cdm_source_name)
+
+  dplyr::collect(cdm$cdm_source) %>%
+    dplyr::mutate(vocabulary_version = dplyr::coalesce(.env$vocab_version, .data$vocabulary_version)) %>%
+    dplyr::mutate(person_cnt = .env$person_cnt, observation_period_cnt = .env$observation_period_cnt) %>%
+    dplyr::select("cdm_source_name",
+                  "cdm_version",
+                  "cdm_holder",
+                  "cdm_release_date",
+                  "vocabulary_version",
+                  "person_cnt",
+                  "observation_period_cnt") %>%
+    as.list() %>%
+    magrittr::set_class("cdm_snapshot")
+}
+
+print.cdm_snapshot <- function(x, ...) {
+  cli::cat_rule(x$cdm_source_name)
+  purrr::walk2(names(x[-1]), x[-1], ~cli::cat_bullet(.x, ": ", .y))
+}
+
+setOldClass("cdm_reference")
+setOldClass("CohortSet")