+++ tests/testthat/test-cdm.R
@@ -1,6 +1,8 @@
 library(testthat)
 library(dplyr, warn.conflicts = FALSE)
 
+### Using DBI drivers ------
+
 test_that("cdm reference works locally", {
   skip_if(Sys.getenv("LOCAL_POSTGRESQL_USER") == "")
   con <- DBI::dbConnect(RPostgres::Postgres(),
@@ -11,7 +13,11 @@ test_that("cdm reference works locally", {
 
   expect_true(is.character(listTables(con, schema = Sys.getenv("LOCAL_POSTGRESQL_CDM_SCHEMA"))))
 
-  cdm <- cdm_from_con(con, cdm_schema = Sys.getenv("LOCAL_POSTGRESQL_CDM_SCHEMA"), cdm_tables = tbl_group("vocab"))
+  cdm <- cdm_from_con(con, cdm_schema = Sys.getenv("LOCAL_POSTGRESQL_CDM_SCHEMA"), cdm_tables = tbl_group("default"))
+
+  expect_error(assert_tables(cdm, "cost"))
+  expect_true(version(cdm) %in% c("5.3", "5.4"))
+  expect_s3_class(snapshot(cdm), "cdm_snapshot")
 
   expect_true(is.null(verify_write_access(con, write_schema = "scratch")))
 
@@ -34,7 +40,12 @@ test_that("cdm reference works on postgres", {
 
   expect_true(is.character(listTables(con, schema = Sys.getenv("CDM5_POSTGRESQL_CDM_SCHEMA"))))
 
+  expect_null(verify_write_access(con, Sys.getenv("CDM5_POSTGRESQL_SCRATCH_SCHEMA")))
+
   cdm <- cdm_from_con(con, cdm_schema = Sys.getenv("CDM5_POSTGRESQL_CDM_SCHEMA"), cdm_tables = tbl_group("vocab"))
+
+  expect_error(assert_tables(cdm, "person"))
+  expect_true(version(cdm) %in% c("5.3", "5.4"))
 
   expect_true("concept" %in% names(cdm))
   expect_s3_class(collect(head(cdm$concept)), "data.frame")
@@ -60,6 +71,9 @@ test_that("cdm reference works on sql server", {
 
   cdm <- cdm_from_con(con, cdm_schema = c("CDMV5", "dbo"), cdm_tables = tbl_group("vocab"))
 
+  expect_error(assert_tables(cdm, "person"))
+  expect_true(version(cdm) %in% c("5.3", "5.4"))
+
   expect_true("concept" %in% names(cdm))
   expect_s3_class(collect(head(cdm$concept)), "data.frame")
 
@@ -67,9 +81,9 @@ test_that("cdm reference works on sql server", {
   expect_null(verify_write_access(con, write_schema = c("tempdb", "dbo")))
 
   cohort <- dplyr::tibble(cohort_id = 1L,
-                           subject_id = 1L:2L,
-                           cohort_start_date = c(Sys.Date(), as.Date("2020-02-03")),
-                           cohort_end_date = c(Sys.Date(), as.Date("2020-11-04")))
+                          subject_id = 1L:2L,
+                          cohort_start_date = c(Sys.Date(), as.Date("2020-02-03")),
+                          cohort_end_date = c(Sys.Date(), as.Date("2020-11-04")))
 
   DBI::dbWriteTable(con, DBI::Id(catalog = "tempdb", schema = "dbo", table = "cohort"), cohort, overwrite = TRUE)
 
@@ -101,27 +115,39 @@ test_that("cdm reference works on redshift", {
                         user     = Sys.getenv("CDM5_REDSHIFT_USER"),
                         password = Sys.getenv("CDM5_REDSHIFT_PASSWORD"))
 
+
+  expect_null(verify_write_access(con, Sys.getenv("CDM5_REDSHIFT_SCRATCH_SCHEMA")))
+
   expect_true(is.character(listTables(con, schema = Sys.getenv("CDM5_REDSHIFT_CDM_SCHEMA"))))
 
-  cdm <- cdm_from_con(con, cdm_schema = Sys.getenv("CDM5_REDSHIFT_CDM_SCHEMA"), cdm_tables = tbl_group("vocab"))
+  cdm <- cdm_from_con(con, cdm_schema = Sys.getenv("CDM5_REDSHIFT_CDM_SCHEMA"), cdm_tables = tbl_group("default"))
+
+  expect_error(assert_tables(cdm, "cost"))
+  expect_true(version(cdm) %in% c("5.3", "5.4"))
 
   expect_true("concept" %in% names(cdm))
   expect_s3_class(collect(head(cdm$concept)), "data.frame")
 
   expect_equal(dbms(con), "redshift")
-  expect_equal(dbms(cdm), "redshift")
 
   DBI::dbDisconnect(con)
 })
 
 
 test_that("cdm reference works on duckdb", {
+  skip_if_not(rlang::is_installed("duckdb", version = "0.6"))
 
   con <- DBI::dbConnect(duckdb::duckdb(), dbdir = eunomia_dir())
 
+  expect_null(verify_write_access(con, "main"))
+
   expect_true(is.character(listTables(con)))
 
-  cdm <- cdm_from_con(con, cdm_tables = tbl_group("vocab"))
+  cdm <- cdm_from_con(con, cdm_tables = tbl_group("default"))
+
+  expect_error(assert_tables(cdm, "cost"))
+  expect_true(version(cdm) %in% c("5.3", "5.4"))
+  expect_s3_class(snapshot(cdm), "cdm_snapshot")
 
   expect_true("concept" %in% names(cdm))
   expect_s3_class(collect(head(cdm$concept)), "data.frame")
@@ -130,6 +156,8 @@ test_that("cdm reference works on duckdb", {
 })
 
 test_that("inclusion of cohort tables", {
+  skip_if_not(rlang::is_installed("duckdb", version = "0.6"))
+
   con <- DBI::dbConnect(duckdb::duckdb(), dbdir = eunomia_dir())
 
   cohort <- dplyr::tibble(cohort_id = 1L,
@@ -157,6 +185,8 @@ test_that("inclusion of cohort tables", {
 })
 
 test_that("collect a cdm", {
+  skip_if_not(rlang::is_installed("duckdb", version = "0.6"))
+
   con <- DBI::dbConnect(duckdb::duckdb(), dbdir = eunomia_dir())
   cdm <- cdm_from_con(con)
 
@@ -176,9 +206,11 @@ test_that("collect a cdm", {
 
 
 test_that("stow and cdm_from_files works", {
+  skip_if_not(rlang::is_installed("duckdb", version = "0.6"))
+
   save_path <- file.path(tempdir(), paste0("tmp_", paste(sample(letters, 10, replace = TRUE), collapse = "")))
   dir.create(save_path)
-  cdm_tables <- c("person", "observation_period")
+  cdm_tables <- c("person", "observation_period", "cdm_source", "vocabulary")
   con <- DBI::dbConnect(duckdb::duckdb(), dbdir = eunomia_dir())
 
   # Test tidyselect in cdm_from_con. Should not produce message about ambiguous names.
@@ -192,25 +224,151 @@ test_that("stow and cdm_from_files works", {
 
   stow(cdm, path = save_path)
 
-  expect_equal(list.files(save_path), c("observation_period.parquet" ,"person.parquet"))
+  expect_setequal(list.files(save_path), paste0(cdm_tables, ".parquet"))
 
-  # test tidyselect in cdm_from_files. Should not produce message about ambiguous names.
-  expect_message(cdm_from_files(save_path, cdm_tables = matches("person|observation_period")), NA)
-  expect_message(cdm_from_files(save_path, cdm_tables = c(person, observation_period)), NA)
-  expect_message(cdm_from_files(save_path, cdm_tables = c("person", "observation_period")), NA)
-  expect_message(cdm_from_files(save_path, cdm_tables = all_of(cdm_tables)), NA)
+  expect_message(cdm_from_files(save_path), NA)
+  expect_warning(cdm_from_files(save_path, cdm_tables = all_of(cdm_tables)), "deprecated")
 
-  local_cdm <- cdm_from_files(save_path, cdm_tables = all_of(cdm_tables))
+  local_cdm <- cdm_from_files(save_path)
   expect_s3_class(local_cdm, "cdm_reference")
   expect_equal(local_cdm$person, collect(cdm$person))
-  expect_output(validate_cdm(local_cdm))
+  expect_s3_class(snapshot(cdm), "cdm_snapshot")
 
-  local_arrow_cdm <- cdm_from_files(save_path, cdm_tables = all_of(cdm_tables), as_data_frame = FALSE)
+  local_arrow_cdm <- cdm_from_files(save_path, as_data_frame = FALSE)
   expect_s3_class(local_arrow_cdm, "cdm_reference")
   expect_equal(collect(local_arrow_cdm$person), collect(cdm$person))
-  expect_output(validate_cdm(local_arrow_cdm))
+  expect_error(validate_cdm(local_arrow_cdm))
 
   DBI::dbDisconnect(con, shutdown = TRUE)
 })
 
 
+## Using DatabaseConnector DBI driver -----
+
+library(testthat)
+library(dplyr, warn.conflicts = FALSE)
+
+test_that("DatabaseConnector cdm reference works on local postgres", {
+  skip_if(Sys.getenv("LOCAL_POSTGRESQL_USER") == "")
+
+  con <- DBI::dbConnect(DatabaseConnector::DatabaseConnectorDriver(),
+                        dbms     = "postgresql",
+                        server   = Sys.getenv("LOCAL_POSTGRESQL_SERVER"),
+                        user     = Sys.getenv("LOCAL_POSTGRESQL_USER"),
+                        password = Sys.getenv("LOCAL_POSTGRESQL_PASSWORD"))
+
+  expect_true(is.character(listTables(con, schema = Sys.getenv("LOCAL_POSTGRESQL_CDM_SCHEMA"))))
+
+  cdm <- cdm_from_con(con, cdm_schema = Sys.getenv("LOCAL_POSTGRESQL_CDM_SCHEMA"), cdm_tables = tbl_group("default"))
+
+  expect_error(assert_tables(cdm, "cost"))
+  expect_true(version(cdm) %in% c("5.3", "5.4"))
+  expect_s3_class(snapshot(cdm), "cdm_snapshot")
+
+  # debugonce(verify_write_access)
+  expect_true(is.null(verify_write_access(con, write_schema = "scratch")))
+
+  expect_true("concept" %in% names(cdm))
+  expect_s3_class(collect(head(cdm$concept)), "data.frame")
+
+  expect_equal(dbms(cdm), "postgresql")
+
+  DBI::dbDisconnect(con)
+})
+
+test_that("DatabaseConnector cdm reference works on postgres", {
+  skip_if(Sys.getenv("CDM5_POSTGRESQL_USER") == "")
+
+  con <- DBI::dbConnect(DatabaseConnector::DatabaseConnectorDriver(),
+                        dbms     = "postgresql",
+                        server   = Sys.getenv("CDM5_POSTGRESQL_SERVER"),
+                        user     = Sys.getenv("CDM5_POSTGRESQL_USER"),
+                        password = Sys.getenv("CDM5_POSTGRESQL_PASSWORD"))
+
+  expect_true(is.character(listTables(con, schema = Sys.getenv("CDM5_POSTGRESQL_CDM_SCHEMA"))))
+
+  cdm <- cdm_from_con(con, cdm_schema = Sys.getenv("CDM5_POSTGRESQL_CDM_SCHEMA"), cdm_tables = tbl_group("default"))
+
+  expect_error(assert_tables(cdm, "cost"))
+  expect_true(version(cdm) %in% c("5.3", "5.4"))
+  expect_s3_class(snapshot(cdm), "cdm_snapshot")
+
+  # debugonce(verify_write_access)
+  expect_true(is.null(verify_write_access(con, write_schema = Sys.getenv("CDM5_POSTGRESQL_SCRATCH_SCHEMA"))))
+
+  expect_true("concept" %in% names(cdm))
+  expect_s3_class(collect(head(cdm$concept)), "data.frame")
+
+  expect_equal(dbms(cdm), "postgresql")
+
+  DBI::dbDisconnect(con)
+})
+
+
+test_that("DatabaseConnector cdm reference works on redshift", {
+  skip_if(Sys.getenv("CDM5_REDSHIFT_USER") == "")
+
+  con <- DBI::dbConnect(DatabaseConnector::DatabaseConnectorDriver(),
+                        dbms     = "redshift",
+                        server   = Sys.getenv("CDM5_REDSHIFT_SERVER"),
+                        user     = Sys.getenv("CDM5_REDSHIFT_USER"),
+                        password = Sys.getenv("CDM5_REDSHIFT_PASSWORD"))
+
+  expect_true(is.character(listTables(con, schema = Sys.getenv("CDM5_POSTGRESQL_CDM_SCHEMA"))))
+
+  cdm <- cdm_from_con(con, cdm_schema = Sys.getenv("CDM5_REDSHIFT_CDM_SCHEMA"), cdm_tables = tbl_group("default"))
+
+  expect_error(assert_tables(cdm, "cost"))
+  expect_true(version(cdm) %in% c("5.3", "5.4"))
+  expect_s3_class(snapshot(cdm), "cdm_snapshot")
+
+  # debugonce(verify_write_access)
+  expect_true(is.null(verify_write_access(con, write_schema = Sys.getenv("CDM5_REDSHIFT_SCRATCH_SCHEMA"))))
+
+  expect_true("concept" %in% names(cdm))
+  expect_s3_class(collect(head(cdm$concept)), "data.frame")
+
+  expect_equal(dbms(cdm), "redshift")
+
+  DBI::dbDisconnect(con)
+})
+
+
+test_that("DatabaseConnector cdm reference works on sql server", {
+  skip_if(Sys.getenv("CDM5_SQL_SERVER_USER") == "")
+  # skip("DatabaseConnector does not preserve logical datatypes")
+  # skip("sql server test database cdm5.dbo.person does not have birth_datetime")
+
+  con <- DBI::dbConnect(DatabaseConnector::DatabaseConnectorDriver(),
+                        dbms     = "sql server",
+                        server   = Sys.getenv("CDM5_SQL_SERVER_SERVER"),
+                        user     = Sys.getenv("CDM5_SQL_SERVER_USER"),
+                        password = Sys.getenv("CDM5_SQL_SERVER_PASSWORD"))
+
+  cdm_schema <- strsplit(Sys.getenv("CDM5_SQL_SERVER_CDM_SCHEMA"), "\\.")[[1]]
+  expect_true(is.character(listTables(con, schema = cdm_schema)))
+
+  cdm <- cdm_from_con(con, cdm_schema = cdm_schema, cdm_tables =  c("cdm_source", "person", "observation_period", "vocabulary"))
+
+  expect_error(assert_tables(cdm, "cost"))
+  expect_true(version(cdm) %in% c("5.3", "5.4"))
+  # expect_s3_class(snapshot(cdm), "cdm_snapshot")
+  # df <- DBI::dbGetQuery(con, "select * from cdmv5.dbo.person")
+
+  expect_true(is.null(verify_write_access(con, write_schema = Sys.getenv("CDM5_SQL_SERVER_SCRATCH_SCHEMA"))))
+
+  expect_true("person" %in% names(cdm))
+  expect_s3_class(collect(head(cdm$person)), "data.frame")
+
+  expect_equal(dbms(cdm), "sql server")
+
+  DBI::dbDisconnect(con)
+})
+
+test_that("autodetect cdm version works", {
+  skip_if_not(rlang::is_installed("duckdb", version = "0.6"))
+  con <- DBI::dbConnect(duckdb::duckdb(), dbdir = eunomia_dir())
+  cdm <- cdm_from_con(con, cdm_tables = tbl_group("default"), cdm_version = "auto")
+  expect_true(version(cdm) == c("5.3"))
+  DBI::dbDisconnect(con, shutdown = TRUE)
+})